classification/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.26.0
pillow>=10.0.0
pyyaml>=6.0.0

# CLIP and vision models
open-clip-torch>=2.20.0
timm>=0.9.0
ftfy>=6.1.0
regex>=2023.0.0

# Data loading
webdataset>=0.2.0
einops>=0.7.0

# Utilities
tqdm>=4.66.0
pandas>=2.0.0
scipy>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
yacs>=0.1.8
iopath>=0.1.0

# Configuration
portalocker>=2.8.0
huggingface-hub>=0.20.0
safetensors>=0.4.0
fsspec>=2024.0.0

# Optional but recommended
gdown>=5.0.0
beautifulsoup4>=4.12.0
